更新售后的金额展示

+++ moss-business/src/main/java/com/ycandyz/master/model/mall/MallAfterSalesVO.java
@@ -120,7 +120,7 @@ public class MallAfterSalesVO {
     @ApiModelProperty(value = "关联订单详情中商品信息表")
     private MallItemByMallOrderDetailVO itemInfo;
     @ApiModelProperty(value = "总计金额")
-    private Integer allMoney;
+    private BigDecimal allMoney;
     @ApiModelProperty(value = "关联买家快递表")
     private MallBuyerShippingVO buyerShipping;
     @ApiModelProperty(value = "关联买家寄出快递日志表")
